Fix bugs in trace() with a interpolating method
In this commit, I applied an interpolating method which perform
the texture Y calcaulation directly in `renderer.cpp`, so the
variable `textureStep` is never need.

There are two main problems in original code:
1. The `screenY` overflow when the wall is too close with player,
consequently leading to the strange effect in the frame.
2. There are some noise at the lower part of the frame due to the
accumulated float number error.

In `raycaster.h`, I add the following definition to indicate the
size of texture.
- TEXTURE_XS	: the exponential part of the texture size
- TEXTURE_SIZE	: automatically generated by TEXTURE_XS