Allow multiple kinds of obstacles
Create a new definition of map storing format, which allows multiple
inds of obstacles represented in it.

The current setting:
A 32 x 32 wide square map, an obstacle use 4 bits, and using uint32_t
to store the map.

Most of the old features are still alive, using an odd value of hit
direction to indicate a vertical hit wall, otherwise, it is a horizontal
hit one, in addition, the value of hitDirection/2 indicating the
textureNo.